requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
kiwisolver==1.4.8
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.2.0
platformdirs==4.3.7
pycuda==2025.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytools==2025.1.2
six==1.17.0
typing_extensions==4.13.0
